+++ meta_package_manager/tests/test_pool.py
@@ -27,7 +27,7 @@ from ..pool import pool
 
 def test_manager_count():
     """Check all implemented package managers are accounted for, and unique."""
-    assert len(pool) == 27
+    assert len(pool) == 28
     assert len(pool) == len(pool.all_manager_ids)
     assert pool.all_manager_ids == tuple(sorted(set(pool)))